move categorical, categorical_single_column to controls
