+++ tests/controls/test_explore.py
@@ -5,12 +5,12 @@ import pytest
 
 import seismometer.controls.explore as undertest
 from seismometer import seismogram
-from seismometer.data.performance import MetricGenerator
-from seismometer.plot.mpl.categorical import CategoricalOptionsWidget, ExploreCategoricalPlots
-from seismometer.plot.mpl.categorical_single_column import (
+from seismometer.controls.categorical import CategoricalOptionsWidget, ExploreCategoricalPlots
+from seismometer.controls.categorical_single_column import (
     CategoricalFeedbackSingleColumnOptionsWidget,
     ExploreSingleCategoricalPlots,
 )
+from seismometer.data.performance import MetricGenerator
 from seismometer.table.analytics_table import AnalyticsTableOptionsWidget, ExploreBinaryModelAnalytics
 
 
@@ -1154,7 +1154,7 @@ class TestExploreCategoricalPlots:
         assert widget.disabled is False
 
         expected_plot_code = (
-            "seismometer.plot.mpl.categorical.ordinal_categorical_plot(['Metric1', 'Metric2'], "
+            "seismometer.controls.categorical.ordinal_categorical_plot(['Metric1', 'Metric2'], "
             + "{'C1': ['C1.1', 'C1.2'], 'C2': ['C2.1', 'C2.2']}, title='Unit Test Title')"
         )
         assert widget.current_plot_code == expected_plot_code
@@ -1263,7 +1263,7 @@ class TestExploreSingleCategoricalPlots:
         assert widget.disabled is False
 
         expected_plot_code = (
-            "seismometer.plot.mpl.categorical_single_column.ordinal_categorical_single_col_plot"
+            "seismometer.controls.categorical_single_column.ordinal_categorical_single_col_plot"
             + "('Metric1', {'Age': ('20-30', '30-40')}, title='Unit Test Title')"
         )
         assert widget.current_plot_code == expected_plot_code
